Update sortphotos.py
In the `execute` function, `os.read(fd, 4096)` has a bug when dealing with multi-byte characters. If this character is precisely split into two byte segments, UTF-8 decoding will report an error.
Optimization method: Use try-except to catch the position of decoding errors, and assign the remaining bytes starting from the error position to the leftover variable. Then, during the decoding of the next byte segment, concatenate and decode it.